Refactor: remove shared `PostProps` and `createPostSchema`, redefine types for `notion2md` and `astro` independently
- Removed `types/content.ts` and decoupled `Astro` dependencies from `notion2md`.
- Inlined `PostProps` definition directly in `notion2md/src/types.ts`.
- Reimplemented `schema` logic for `astro:content` with `zod` directly in the config.
- Updated `notion2md` startup script to include build step in `package.json`.
