+++ src/content/config.ts
@@ -1,9 +1,25 @@
-import { defineCollection } from "astro:content";
-import { createPostSchema } from "../../types/content.js";
+import { defineCollection, z } from "astro:content";
 
 const postsCollection = defineCollection({
   type: "content",
-  schema: createPostSchema,
+  schema: ({ image }) =>
+    z.object({
+      draft: z.boolean(),
+      title: z.string(),
+      slug: z.string().optional(), // Astro가 폴더명을 slug로 사용하므로 optional로 변경
+      category: z.string().nullable(),
+      tags: z.array(z.string()).transform((tags) => {
+        // 공백을 하이픈으로 변환 후 중복 제거
+        const transformedTags = tags.map((tag) => tag.replace(/\s+/g, "-"));
+        return [...new Set(transformedTags)];
+      }),
+      summary: z.string(),
+      author: z.array(z.string()),
+      heroImage: image().nullable(),
+      publishedAt: z.string().nullable(),
+      createdAt: z.coerce.date(),
+      updatedAt: z.coerce.date().nullable(),
+    }),
 });
 
 export const collections = {
